"게시물 기능 추가"

+++ src/main/java/com/bitacademy/myportal/service/UserService.java
@@ -6,5 +6,5 @@ public interface UserService {
 	public boolean join(UserVo vo);	//	회원 가입
 	public UserVo getUser(String email, String password);	//	로그인 메서드
 	public UserVo getUser(String email);	//	 중복 로그인 체크용
-	public Long updateUser(UserVo vo); // 회원 정보 수정 
+	public int updateUser(UserVo vo); // 회원 정보 수정 
 }
